log/slog is now the default logger
