--- logging/kit/kit_suite_test.go
@@ -1,13 +1,0 @@
-package goakit_test
-
-import (
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-
-	"testing"
-)
-
-func TestKit(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Goakit Suite")
-}
