--- logging/slog/logrus_suite_test.go
@@ -1,13 +1,0 @@
-package goaslog_test
-
-import (
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-
-	"testing"
-)
-
-func TestSlog(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Slog Suite")
-}
